agent_template_83093/requirements.txt
```
python-dotenv
langchain-nvidia-ai-endpoints
langgraph
langchain-core
langchain-community
```